API Reference

Public

Internal

BigBed.Reader — Method

BigBed.Reader(input::IO)

Create a reader for bigBed file format.

Note that input must be seekable.

BigBed.Record — Method

BigBed.Record()

Create an unfilled bigBed record.

BigBed.Writer — Method

BigBed.Writer(output::IO, chromlist; binsize=64)

Create a data writer of the bigBed file format.

Arguments

output: data sink

chromlist: chromosome list with length

binsize=64: size of a zoom with the highest resolution

Examples

output = open("data.bb", "w")
writer = BigBed.Writer(output, [("chr1", 12345), ("chr2", 9100)])
write(writer, ("chr1", 101, 150, "gene 1"))
write(writer, ("chr2", 211, 250, "gene 2"))
close(writer)

BigBed.blockcount — Method

blockcount(record::Record)::Int

Get the number of blocks (exons) in record.

BigBed.blocksizes — Method

blocksizes(record::Record)::Vector{Int}

Get the block (exon) sizes of record.

BigBed.blockstarts — Method

blockstarts(record::Record)::Vector{Int}

Get the block (exon) starts of record.

Note that the first base is numbered 1.

BigBed.chrom — Method

chrom(record::Record)::String

Get the chromosome name of record.

BigBed.chromend — Method

chromend(record::Record)::Int

Get the end position of record.

BigBed.chromid — Method

chromid(record::Record)::UInt32

Get the chromosome ID of record.

BigBed.chromlist — Method

chromlist(reader::BigBed.Reader)::Vector{Tuple{String,Int}}

Get the (name, length) pairs of chromosomes/contigs.

BigBed.chromstart — Method

chromstart(record::Record)::Int

Get the start position of record.

BigBed.itemrgb — Method

itemrgb(record::Record)::ColorTypes.RGB

Get the RGB value of record.

The return type is defined in ColorTypes.jl.

BigBed.name — Method

name(record::Record)::String

Get the name of record.

BigBed.optionals — Method

optionals(record::Record)::Vector{String}

Get optional fields as strings.

BigBed.score — Method

score(record::Record)::Int

Get the score between 0 and 1000.

BigBed.strand — Method

strand(record::Record)::GenomicFeatures.Strand

Get the strand of record.

BigBed.thickend — Method

thickend(record::Record)::Int

Get the end position at which record is drawn thickly.

BigBed.thickstart — Method

thickstart(record::Record)::Int

Get the starting position at which record is drawn thickly.

Note that the first base is numbered 1.